refactor(api): add new fields for due, surcharge, and surcharge percentage
Added new fields for due, surcharge, and surcharge percentage to the Course model. Also, renamed files according to conventions.
